fixing this fucking race condition once and for all
previously, we were getting an embedded image's dimensions in
img2c_task(), but unfortunately that doesn't get run if the image hasn't
changed! so, in cases where the stylesheet had changed but the image
hadn't, the dimensions would be set to `0x0`, and no texture would show
up.

bummer.

it's fixed now though.