stylequad: moving away from querying our own props
currently, stylequad queries tons of properties on the parent element,
but that's really none of the stylequad's business. so, we're moving
toward a more push-based API, where the element queries its own
properties and assigns them to the stylequad. this is also in
preparation for having several stylequads in the same element.