cocoa: fix blank window under El Capitan
there is a bug in Cocoa which manifests when opening a window which is
too large for the screen and must be immediately resized. when getting
the setFrame message, convertSizeToBacking will incorrectly convert GL
pixels to physical pixels, which on Retina displays will be some
multiple of GL pixels.

as it turns out, apple says we only need to do the coordinate system
conversion if we've specified wantsBestResolutionOpenGLSurface. we don't
currently, so we don't need it at all.

thanks, apple.

thapple.