platform/x11-xcb: remove GLX_OML_sync_control code
nothing on Linux has given me as much trouble as this FPS locking code.
it breaks on nvidia, or it breaks with certain versions of Mesa.

whatever. hardcoding 60fps like I do on Mac and Windows for now. will
move to a better way of determining vblank interval in the future. EGL
maybe? need to investigate.