element: implement RTB_ELEM_ACTIVE_DURING_DRAG
usually, when the user clicks and then drags outside the bounds of an
element, the element turns back to the "normal" state. this is natural
for e.g. buttons, where the click event only fires when the mouse-up
occurs within the bounds.

however, for controls like value elements, where the drag itself is a
continuous adjustment, it makes sense that the element should stay in
the active state. setting RTB_ELEM_ACTIVE_DURING_DRAG on an element now
does this.