first attempt at changing window scale while the window is open
actually works pretty well! there's one major thing i want to try and
address in the future:

it would be nice to not have to delete all the texture fonts and
recreate them from scratch. i think this malloc() thrashing is part of
the perf hit when changing DPI and it would be nice to not have to do
it.

going to take a bit of surgery in texture-font to deal with, and will
have to lift out the inner glyph render/cache loop, but it's doable.